cleaned up garbage code, removed bugs
ha ha, i thought some decimal was hex so there was a bug where the sha
would stop if there was a 2 ever! wacky! but it's fixed now. a legacy
param in writeppm was removed, because it did nothing.

removed a few useless functions, added some distance stuff, etc etc